added upper south platte and blue river basin get routes, fixed broken functionality on index.js
