+++ src/client/js/main.js
@@ -1,15 +1,16 @@
 (function () {
   // console.log('sanity check!');
 
-  $('[name="river-basin"]').on('change', function() {
-    console.log('changed!');
-    var riverName = $('[name="river-basin"]').val();
-    console.log(riverName)
+//select river basin, and this auto populates the sites dropdown
 
+  $('[name="river-basin"]').on('change', function() {
+    var riverName = $('[name="river-basin"]').val();
     $.ajax({
       type: 'GET',
       url: '/sites?river='+riverName,
       success: function(data) {
+        let dropdown = $('#site-dropdown');
+        dropdown.empty();
         $.each(data.sites, function(k,v) {
           var site = v["site_name"]
           $('#site-dropdown').append($('<option></option>')
@@ -17,7 +18,7 @@
         });
       },
       error: function(data) {
-        console.log('error');
+        return error;
       }
     });
   });
